app/requirements.txt
setuptools
flask
flask-cors
python-dotenv
flasgger
rdkit
scaffoldgraph[rdkit]
pandas